Refactor seedu.duke packages into findoursep
