+++ build.gradle
@@ -33,7 +33,7 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.FindOurSEP")
+    mainClass.set("findoursep.FindOurSEP")
 }
 
 shadowJar {
