+++ src/test/java/findoursep/FindOurSEPTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package findoursep;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class FindOurSEPTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
